[REF] Group tests into positive and negative
Since we expected a different behaviour from positive and negative tests, the
tests got grouped, and now it's execution is more robust i.e. all negative
tests will be executed if one of the operation fails as expected.

Also the output has been now standardized, so that a test always reports
success or failure independently of its type.